dashboard and staff front end added

+++ Web App/edi_project/lib/FireBaseHelper.dart
@@ -6,32 +6,57 @@ import 'package:firebase_storage/firebase_storage.dart';
 class FireBaseHelper{
   FirebaseFirestore _db = FirebaseFirestore.instance; 
 
-  getAllStudents() async{
-    QuerySnapshot temp = await _db.collection("users").get();
+  getAllStudents(int flag) async{
+    QuerySnapshot temp;
+    if (flag == 1)
+      temp = await _db.collection("users").get();
+
+    if (flag == 2) 
+      temp = await _db.collection("staff").get();
+
     return temp;
   }
 
-  getAllDefaulters() async{
-    QuerySnapshot temp = await _db.collection("defaulters").get();
+  getAllDefaulters(int flag) async{
+     QuerySnapshot temp ;
+    if (flag == 1)
+      temp = await _db.collection("defaulters").get();
+    if (flag == 2)
+      temp = await _db.collection("defaulter_staff").get();
     return temp;
   }
 
-  deleteStudentDatabase(String grno) async{
+  deleteStudentDatabase(String grno, int flag) async{
     String path = "Photos/" + grno + "/" + grno;
-    await _db.collection("users").doc(grno).delete();
+    
+    if (flag == 1)
+      await _db.collection("users").doc(grno).delete();
+    if (flag == 2){
+      print ("in");
+      await _db.collection("staff").doc(grno).delete();}
+    
     await FirebaseStorage.instance.ref().child(path).delete();
   }
 
-  addNewStudent(StudentModel student) async{
-    await _db.collection("users").doc(student.grno).set(student.toMap());
+  addNewStudent(student, int flag) async{
+    if (flag == 1)
+      await _db.collection("users").doc(student.grno).set(student.toMap());
+
+    if (flag == 2)
+      await _db.collection("staff").doc(student.grno).set(student.toMap());
   }
 
   deleteDefaulterStudent(String grno) async{
     await _db.collection("defaulters").doc(grno).delete();
   }
 
-  getStudentInfo(String grno) async {
-    var temp = _db.collection("users").doc(grno).get();
+  getStudentInfo(String grno, int flag) async {
+    var temp;
+    if (flag == 1)
+      temp = _db.collection("users").doc(grno).get();
+
+    if (flag == 2)
+      temp = _db.collection("staff").doc(grno).get();
     return temp;
   }
 }
